+++ changelog.md
@@ -7,6 +7,10 @@ SPDX-License-Identifier: AGPL-3.0-only
 
 ← [Back to README](readme.md)
 
+## v8.6.6 (August 22, 2026)
+- Add traslated manuals.
+- Remove Gemini interactive guide link.
+
 ## v8.6.5 (August 21, 2026)
 - Add message ranges to job queue display.